server/api/api/requirements/base.txt
python-dotenv==1.0.0
django==4.2.6
django-jsonform==2.19.1
djangorestframework==3.14.0
drf-writable-nested==0.7.0
djoser==2.2.0
Pillow==10.0.1
django-ckeditor==6.7.0
psycopg[binary]==3.1.12
django-cities-light==3.10.0